refactor(Select): remove the interface to index file
