+++ src/components/inputs/Select/index.tsx
@@ -1,8 +1,34 @@
 import { useState, useRef, useEffect } from "react";
 
 import { SelectUI } from "./interface";
-import { ISelectProps } from "./interfaces/Select.interface";
-import { states } from "./types/Select.States.type";
+import { States, states } from "./types/Select.States.type";
+import { Size } from "./types/Select.Size.type";
+
+export interface ISelectOptions {
+  id: string;
+  label: string;
+  isDisabled: boolean;
+}
+
+export interface ISelectProps {
+  label?: string;
+  name?: string;
+  id: string;
+  placeholder?: string;
+  isDisabled?: boolean;
+  value?: string | number;
+  isRequired?: boolean;
+  state?: States;
+  errorMessage?: string;
+  validMessage?: string;
+  size?: Size;
+  isFullWidth?: boolean;
+  options: ISelectOptions[];
+  handleChange?: (event: MouseEvent) => void;
+  handleFocus?: (event: FocusEvent) => void;
+  handleBlur?: (event: FocusEvent) => void;
+  handleClick?: (event: MouseEvent) => void;
+}
 
 const defaultIsDisabled = false;
 const defaultIsRequired = false;
