--- src/components/inputs/Select/interfaces/Select.Options.interface.ts
@@ -1,5 +1,0 @@
-export interface ISelectOptions {
-  id: string;
-  label: string;
-  isDisabled: boolean;
-}
