--- src/components/inputs/Select/interfaces/Select.Success.interface.ts
@@ -1,6 +1,0 @@
-export interface ISelectStateProps {
-  isDisabled: boolean;
-  state: string;
-  validMessage?: string;
-  errorMessage?: string;
-}
