--- src/components/inputs/Select/interfaces/SelectInterface.interface.ts
@@ -1,7 +1,0 @@
-import { ISelectProps } from "./Select.interface";
-
-export interface ISelectInterfaceProps extends ISelectProps {
-  isFocused?: boolean;
-  openOptions: boolean;
-  onCloseOptions: () => void;
-}
