+++ src/components/inputs/Select/stories/Select.form.Controller.tsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { Select } from "../index";
 import { StyledForm } from "./styles";
 import { Button } from "../../Button";
-import { ISelectProps } from "../interfaces/Select.interface";
+import { ISelectProps } from "..";
 
 const InForm = (props: ISelectProps) => {
   const { value = "", state = "pending", isRequired } = props;
