+++ src/components/inputs/Select/stories/SelectController.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Select } from "..";
-import { ISelectProps } from "../interfaces/Select.interface";
+import { ISelectProps } from "..";
 
 const SelectController = (props: ISelectProps) => {
   const { value = "", state = "pending" } = props;
