+++ src/components/inputs/Select/styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { ISelectInterfaceProps } from "./interfaces/SelectInterface.interface";
+import { ISelectInterfaceProps } from "./interface";
 import { colors } from "../../../shared/colors/colors";
 import { typography } from "../../../shared/typography/typography";
